[Fix] #49 Recyclerview Adapter 분리
