+++ app/src/main/java/kr/co/lion/unipiece/ui/infomation/InfoAllActivity.kt
@@ -16,6 +16,10 @@ import kr.co.lion.unipiece.util.setMenuIconColor
 class InfoAllActivity : AppCompatActivity() {
 
     lateinit var activityInfoAllBinding: ActivityInfoAllBinding
+
+    val infoAllAdapter:InfoAllAdapter by lazy {
+        InfoAllAdapter()
+    }
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         activityInfoAllBinding = ActivityInfoAllBinding.inflate(layoutInflater)
@@ -52,45 +56,8 @@ class InfoAllActivity : AppCompatActivity() {
     private fun settingAdapter(){
         activityInfoAllBinding.apply {
             recyclerviewInfoAll.apply {
-                adapter = RecyclerviewAdapter()
+                adapter = infoAllAdapter
                 layoutManager = LinearLayoutManager(this@InfoAllActivity)
-            }
-        }
-    }
-
-
-    inner class RecyclerviewAdapter:RecyclerView.Adapter<RecyclerviewAdapter.ViewHolderClass>(){
-
-        inner class ViewHolderClass(infoAllBinding: InfoAllBinding):RecyclerView.ViewHolder(infoAllBinding.root){
-            var infoAllBinding :InfoAllBinding
-
-            init {
-                this.infoAllBinding = infoAllBinding
-                this.infoAllBinding.root.layoutParams = ViewGroup.LayoutParams(
-                    ViewGroup.LayoutParams.MATCH_PARENT,
-                    ViewGroup.LayoutParams.WRAP_CONTENT
-                )
-
-            }
-        }
-
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolderClass {
-            var infoAllBinding = InfoAllBinding.inflate(layoutInflater)
-            var viewHolder = ViewHolderClass(infoAllBinding)
-            return viewHolder
-        }
-
-        override fun getItemCount(): Int {
-            return 30
-        }
-
-        override fun onBindViewHolder(holder: ViewHolderClass, position: Int) {
-            holder.infoAllBinding.imageView2.setImageResource(R.drawable.icon)
-            holder.infoAllBinding.textInfoAllTitle.text = "UniPiece"
-            holder.infoAllBinding.textInfoAllDate.text = "2024-04-01 ~ 2024-04-26"
-            holder.infoAllBinding.textInfoAllAuthorName.text = "멋쟁이 사람들"
-            holder.infoAllBinding.root.setOnClickListener {
-                startActivity(Intent(this@InfoAllActivity, InfoOneActivity::class.java))
             }
         }
     }
